RootMeasurementWriter also truncates it's output

+++ Examples/Io/Root/include/ActsExamples/Io/Root/RootMeasurementWriter.hpp
@@ -68,7 +68,7 @@ class RootMeasurementWriter final : public WriterT<MeasurementContainer> {
 
     TTree* tree = nullptr;
     // Identification parameters
-    int eventNr;
+    size_t eventNr;
     int volumeID;
     int layerID;
     int surfaceID;
@@ -172,7 +172,7 @@ class RootMeasurementWriter final : public WriterT<MeasurementContainer> {
     ///
     /// @param eventNr The event number
     /// @param geoID The geometry identifier of the measurement
-    void fillIdentification(int evnt, Acts::GeometryIdentifier geoId) {
+    void fillIdentification(size_t evnt, Acts::GeometryIdentifier geoId) {
       eventNr = evnt;
       volumeID = geoId.volume();
       layerID = geoId.layer();
